refactored mapping to ec2instances lib (py-vm-instance-aws) for VMs and sessions
