+++ pyproject.toml
@@ -17,7 +17,8 @@ dependencies = [
     "psutil >= 5.9.0",
     "rich >= 12.0.1",
     "sshconf >= 0.2.3",
-    "pycryptodome >= 3.14.1"
+    "pycryptodome >= 3.14.1",
+    "ec2map >= 1.0.0"
 ]
 
 
@@ -34,7 +35,7 @@ dev = [
     "isort >=5.9.3",
     "autoflake >=1.4",
     "flake8 >=4.0.1",
-    "pre-commit >=2.17.0"
+    "pre-commit >=2.17.0",
     ]
 test = [
     "pytest >=5.4.3",
