+++ src/pyclvm/instance/start.py
@@ -16,7 +16,7 @@ def _start_instance(instance_name: str, instance: Instance) -> Any:
         "pending": partial(_in_transition, instance_name, instance),
         "shutting-down": partial(_in_transition, instance_name, instance),
         "rebooting": partial(_in_transition, instance_name, instance),
-    }[instance.state.Name]()
+    }[instance.state.name]()
 
 
 def _is_running(instance_name: str) -> None:
@@ -26,7 +26,6 @@ def _is_running(instance_name: str) -> None:
 def _is_stopped_or_terminated(instance_name: str, instance: Instance) -> None:
     print(f"Starting {instance_name} ...")
     instance.start()
-    instance.wait_until_running()
     print(f"{instance_name} is running")
 
 
@@ -35,7 +34,7 @@ def _in_transition(instance_name: str, instance: Instance) -> None:
         f"{instance_name} is now in transition state. Wait untill current state is determined."
     )
     instance.wait_until_exists()
-    print(f"{instance_name} is {instance.state.Name}")
+    print(f"{instance_name} is {instance.state.name}")
 
 
 def start(*instance_names: str, **kwargs: str) -> Optional[Tuple[Session, str]]:
