+++ src/pyclvm/redirect/stop.py
@@ -1,6 +1,7 @@
 import contextlib
 
-from pyclvm._common.user_data import fetch, remove
+from ec2instances.common.user_data import fetch, remove
+
 from pyclvm.instance import stop as stop_instance
 from pyclvm.ssm.session import stop as terminate_session
