+++ src/pyclvm/ssm/shell.py
@@ -1,7 +1,6 @@
-import time
 from typing import Any, Tuple, Union
 
-from pyclvm.instance import start
+from ec2instances.ec2_instance_mapping import Ec2RemoteShellMapping
 
 
 def shell(
@@ -21,34 +20,5 @@ def shell(
         standard output, standard err if wait=True (default), otherwise ssm_client, command_id, instance_id
 
     """
-    session, instance_id = start(instance_name, **kwargs)
-    ssm_client = session.client("ssm")
-    result = ssm_client.send_command(
-        InstanceIds=[instance_id],
-        DocumentName="AWS-RunShellScript",
-        Parameters={"commands": ["source /etc/bashrc", *commands]},
-    )
-    command_id = result.Command.CommandId
-    # see https://stackoverflow.com/questions/50067035/retrieving-command-invocation-in-aws-ssm
-    time.sleep(2)
-    if not kwargs.get("wait", True):
-        return ssm_client, command_id, instance_id
-    waiter = ssm_client.get_waiter("command_executed")
-    try:
-        waiter.wait(
-            CommandId=command_id,
-            InstanceId=instance_id,
-            WaiterConfig={
-                "Delay": kwargs.get("delay", 5),
-                "MaxAttempts": kwargs.get("attempts", 20),
-            },
-        )
-    finally:
-        result = ssm_client.get_command_invocation(
-            CommandId=command_id,
-            InstanceId=instance_id,
-            PluginName="aws:RunShellScript",
-        )
-        print(result.StandardOutputContent)
-        print(result.StandardErrorContent)
-    return result.StandardOutputContent, result.StandardErrorContent
+    instance = Ec2RemoteShellMapping(**kwargs)[instance_name]
+    return instance.execute(*commands)
